restroom_app/backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
filelock==3.13.1
pipenv==2023.10.24
platformdirs==3.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
virtualenv==20.24.6
